Removed Moja Aplikacja and merged codebase
