--- moja-aplikacja/src/components/IntroContent.js
@@ -1,11 +1,0 @@
-import React from 'react';
-const IntroContent = () => {
-    return (
-        <div>
-            <p>Tutaj powinien być przycisk "Create new character"</p>
-            <p>Tutaj powinien być przycisk "Choose character"</p>
-        </div>
-        
-    );
-};
-export default IntroContent;
