--- moja-aplikacja/src/components/IntroHeader.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import logo from '../images/Logotype.svg';
-const IntroHeader = () => {
-return (
-    <header>
-        <img src={logo} alt="Aplication Logo" height="92" width="247"/>
-    </header>
-    );
-};
-export default IntroHeader;
