--- moja-aplikacja/src/components/Layout.js
@@ -1,9 +1,0 @@
-const Layout = ({children}) => {
-  return (
-    <div>
-      {children}
-    </div>
-  )
-}
-
-export default Layout;
