--- moja-aplikacja/src/components/TopBar.js
@@ -1,9 +1,0 @@
-import React from 'react';
-const TopBar = () => {
-    return (
-      <div>
-      </div>
-    )
-  }
-  
-  export default TopBar;
