+++ src/shared/components/IntroContent.tsx
@@ -1,0 +1,9 @@
+const IntroContent = () => {
+  return (
+    <div>
+      <p>Tutaj powinien być przycisk Create new character</p>
+      <p>Tutaj powinien być przycisk Choose character</p>
+    </div>
+  );
+};
+export default IntroContent;
