+++ src/shared/components/Layout.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+const Layout = ({ children }: { children: ReactNode }) => {
+  return <div>{children}</div>;
+};
+
+export default Layout;
