+++ src/shared/components/TopBar.tsx
@@ -1,0 +1,5 @@
+const TopBar = () => {
+  return <div>Top Bar</div>;
+};
+
+export default TopBar;
